+++ Api/Setup/ServicesRegister.cs
@@ -3,7 +3,7 @@ using Infrastructure.Redis;
 using Microsoft.EntityFrameworkCore;
 using Streaming.DependencyInjection;
 
-namespace Api;
+namespace Api.Setup;
 
 public static class ServicesRegister
 {
